Adds preliminary support for sql sources using dplyr
This adds import.sql, import.pg, import.mysql, and import.sqlite. They
are all aliases to dplyr's src_* family of database functions and
require:

* The arguments used to connect to a database source, such as host,
port, username, and password for non-sqlite and file path for sqlite.

* A `table` argument that specifies the table to collect into a local
data.frame.

In the future, we can remove the `from = table` and replace with
something like `sql(...)` tbl() will allow an arbitrary sql statement
as a string.