Import with S3 classes as described in #42
This commit is an initial attempt at refactoring `import` to use method dispatch rather than a large `switch()` statement.

This is accomplished by creating a new generic function `.import` and implementing each call that was in the `switch` statement as an S3 method. Class names follow the pattern `importfile_*`, with file extensions in the place of the wild card `*`. `import.default` generates the same error message that the switch statement did previously when it failed to recognize a file extension. This is largely to keep tests passing and for consistency, although a future commit may modify this test since method dispatch allows us to specify the unrecognized file format in the error message.

Additional comments on implementation:

* `import.delim` is moved to the top of the file. It is a helper function not called by having a file type of "delim". I would recommend making this `import_delim` in a future release so as to not be confused with S3 methods of `import`.
* All S3 methods are listed in the order of the previous `switch` statement within the `import` file. `import` and it's generic function `.import` are at the bottom of the file. Some projects choose to make the S3 methods separate files. It may be worth considering, but it also generates a lot of overhead for some simple methods.
* All testthat tests are passing. It's worth noting that the current test suite has side effects by generating files in the `tests/testthat` directory. These should be cleaned up, otherwise they cause subsequent tests to fail. I have filed a new issue #68.