Fix failing test for get_type
## Why?
`get_type()` no longer raises an error when the type is not returned. Instead, the extension/format passed to `get_type()` is returned with a message. Instead, `.import.default()` and `.export.default()` raise the error.

## How does this commit address this need?

This changes back the element returned by `get_data()` in the unmatched case to `fmt`, and changes the tests to `expect_message()` and make sure unmatched are passed through untouched.

Additionally, I added my email address to the contributors.